requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django~=4.2
django-cors-headers==4.1.0
django-import-export==3.2.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0

PyMySQL~=1.0.2